+++ pyairtable/models/audit.py
@@ -1,9 +1,10 @@
 from datetime import datetime
 from typing import Any, Dict, List, Optional
 
+import pydantic
 from typing_extensions import TypeAlias
 
-from pyairtable.models._base import AirtableModel, update_forward_refs
+from pyairtable.models._base import AirtableModel, rebuild_models
 
 
 class AuditLogResponse(AirtableModel):
@@ -18,8 +19,8 @@ class AuditLogResponse(AirtableModel):
     pagination: Optional["AuditLogResponse.Pagination"] = None
 
     class Pagination(AirtableModel):
-        next: Optional[str]
-        previous: Optional[str]
+        next: Optional[str] = None
+        previous: Optional[str] = None
 
 
 class AuditLogEvent(AirtableModel):
@@ -28,14 +29,18 @@ class AuditLogEvent(AirtableModel):
 
     See `Audit log events <https://airtable.com/developers/web/api/audit-log-events>`__
     for more information on how to interpret this data structure.
+
+    To avoid namespace conflicts with the Pydantic library, the
+    ``modelId`` and ``modelType`` fields from the Airtable API are
+    represented as fields named ``object_id`` and ``object_type``.
     """
 
     id: str
     timestamp: datetime
     action: str
     actor: "AuditLogActor"
-    model_id: str
-    model_type: str
+    object_id: str = pydantic.Field(alias="modelId")
+    object_type: str = pydantic.Field(alias="modelType")
     payload: "AuditLogPayload"
     payload_version: str
     context: "AuditLogEvent.Context"
@@ -73,4 +78,4 @@ class AuditLogActor(AirtableModel):
 AuditLogPayload: TypeAlias = Dict[str, Any]
 
 
-update_forward_refs(vars())
+rebuild_models(vars())